refactored to give guards to the routes file of each journey
