+++ server/routes/index.ts
@@ -8,8 +8,6 @@ import { JourneyRoutes } from './journeys/routes'
 import { CsipRecordRoutes } from './csip-records/routes'
 import { SearchCsipRoutes } from './manage-csips/routes'
 import { HomePageController } from './controller'
-import journeyStateGuard from '../middleware/journeyStateGuard'
-import config from '../config'
 
 export default function routes(services: Services): Router {
   const router = Router({ mergeParams: true })
@@ -54,10 +52,6 @@ export default function routes(services: Services): Router {
   )
 
   router.use(journeyStateMachine())
-
-  if (config.features.stateGuard || process.env.NODE_ENV === 'e2e-test') {
-    router.use(journeyStateGuard())
-  }
 
   router.use('/:journeyId', JourneyRoutes(services))
   return router
