+++ server/routes/journeys/develop-an-initial-plan/routes.ts
@@ -1,3 +1,4 @@
+import { Request } from 'express'
 import StartJourneyRoutes from './start/routes'
 import { Services } from '../../../services'
 import { JourneyRouter } from '../base/routes'
@@ -12,6 +13,7 @@ import { NextReviewDateRoutes } from './next-review-date/routes'
 import { InterventionDetailsRoutes } from './intervention-details/routes'
 import { ConfirmationRoutes } from './confirmation/routes'
 import { PlanCheckAnswersRoutes } from './check-answers/routes'
+import journeyStateGuard, { JourneyStateGuard, isMissingValues } from '../../../middleware/journeyStateGuard'
 
 function Routes({ csipApiService, auditService }: Services) {
   const { router, get, post } = JourneyRouter()
@@ -36,7 +38,78 @@ export const DevelopPlanRoutes = ({ services, path }: { services: Services; path
   const { router } = JourneyRouter()
 
   router.use('/csip-record/:csipRecordId/develop-an-initial-plan/start', StartJourneyRoutes(services))
+
+  router.use(path, journeyStateGuard(guard))
+
   router.use(path, Routes(services))
 
   return router
+}
+
+const parseIdentifiedNeedIndex = (req: Request) => {
+  const match = req.url.match(/\/(\d+)(?:#[A-Za-z0-9-]+)?$/) || []
+  const index = Number(match[1]) - 1
+  return {
+    success: !Number.isNaN(index) && (req.journeyData.plan!.identifiedNeeds || []).length >= index,
+    isNew: (req.journeyData.plan!.identifiedNeeds || []).length === index,
+    index,
+  }
+}
+
+const guard: JourneyStateGuard = {
+  'identified-needs': (req: Request) => (isMissingValues(req.journeyData.plan!, ['reasonForPlan']) ? '' : undefined),
+  'next-review-date': (req: Request) =>
+    isMissingValues(req.journeyData.plan!, ['identifiedNeeds']) ? '/record-actions-progress/1' : undefined,
+  'summarise-identified-need': (req: Request) =>
+    isMissingValues(req.journeyData.plan!, ['reasonForPlan']) ? '' : undefined,
+  'intervention-details': (req: Request) => {
+    const { success, isNew, index } = parseIdentifiedNeedIndex(req)
+
+    if (!success) {
+      if (!req.journeyData.plan?.reasonForPlan) {
+        return ''
+      }
+
+      if (!req.journeyData.plan?.identifiedNeedSubJourney) {
+        // There's no sub journey present so the intent is not to create a new one
+        return '/identified-needs'
+      }
+
+      // We can figure out if they're in the sub journey and what index they're editing
+      return `/summarise-identified-need/${(req.journeyData.plan?.identifiedNeeds?.length || 0) + 1}`
+    }
+
+    const missing = isMissingValues(
+      isNew ? req.journeyData.plan!.identifiedNeedSubJourney! : req.journeyData.plan!.identifiedNeeds![index]!,
+      ['identifiedNeed'],
+    )
+
+    return missing ? `/summarise-identified-need/${req.url.split('/').pop()}` : undefined
+  },
+  'record-actions-progress': (req: Request) => {
+    const { success, isNew, index } = parseIdentifiedNeedIndex(req)
+
+    if (!success) {
+      if (!req.journeyData.plan?.reasonForPlan) {
+        return ''
+      }
+
+      if (!req.journeyData.plan?.identifiedNeedSubJourney) {
+        // There's no sub journey present so the intent is not to create a new one
+        return '/identified-needs'
+      }
+
+      // We can figure out if they're in the sub journey and what index they're editing
+      return `/intervention-details/${(req.journeyData.plan?.identifiedNeeds?.length || 0) + 1}`
+    }
+
+    const missing = isMissingValues(
+      isNew ? req.journeyData.plan!.identifiedNeedSubJourney! : req.journeyData.plan!.identifiedNeeds![index]!,
+      ['responsiblePerson'],
+    )
+
+    return missing ? `/intervention-details/${req.url.split('/').pop()}` : undefined
+  },
+  'check-answers': (req: Request) =>
+    isMissingValues(req.journeyData.plan!, ['nextCaseReviewDate']) ? '/next-review-date' : undefined,
 }
