+++ server/routes/journeys/referral/routes.ts
@@ -1,3 +1,4 @@
+import { Request } from 'express'
 import { OnBehalfOfRoutes } from './on-behalf-of/routes'
 import { ReferralAreaOfWorkRoutes } from './area-of-work/routes'
 import StartJourneyRoutes from './start/routes'
@@ -16,6 +17,7 @@ import { ReferralCheckAnswersRoutes } from './check-answers/routes'
 import { ConfirmationRoutes } from './confirmation/routes'
 import { Services } from '../../../services'
 import { JourneyRouter } from '../base/routes'
+import journeyStateGuard, { isMissingValues } from '../../../middleware/journeyStateGuard'
 
 function Routes({ csipApiService, auditService }: Services) {
   const { router } = JourneyRouter()
@@ -43,7 +45,55 @@ export const ReferralRoutes = ({ services, path }: { services: Services; path: s
   const { router } = JourneyRouter()
 
   router.use('/prisoners/:prisonerNumber/referral/start', StartJourneyRoutes(services))
+  router.use(path, journeyStateGuard(guard))
   router.use(path, Routes(services))
 
   return router
+}
+
+const redirectAreaOfWorkOrReferrer = (req: Request, page: string) => {
+  const referral = req.journeyData.referral?.onBehalfOfSubJourney || req.journeyData.referral
+
+  if (referral?.isOnBehalfOfReferral === false) {
+    return page === 'area-of-work' ? undefined : '/area-of-work'
+  }
+
+  if (referral?.isOnBehalfOfReferral === true) {
+    return page === 'referrer' ? undefined : '/referrer'
+  }
+
+  return '/on-behalf-of'
+}
+
+const guard = {
+  'area-of-work': (req: Request) => redirectAreaOfWorkOrReferrer(req, 'area-of-work'),
+  referrer: (req: Request) => redirectAreaOfWorkOrReferrer(req, 'referrer'),
+  'proactive-or-reactive': (req: Request) => {
+    if (isMissingValues(req.journeyData.referral!, ['refererArea'])) {
+      return req.journeyData.referral!.onBehalfOfSubJourney?.isOnBehalfOfReferral ? '/referrer' : '/area-of-work'
+    }
+    return undefined
+  },
+  details: (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['isProactiveReferral']) ? '/proactive-or-reactive' : undefined,
+  involvement: (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['incidentLocation', 'incidentType', 'incidentDate'])
+      ? '/details'
+      : undefined,
+  description: (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['incidentInvolvement', 'isStaffAssaulted'])
+      ? '/involvement'
+      : undefined,
+  reasons: (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['descriptionOfConcern']) ? '/description' : undefined,
+  'contributory-factors': (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['knownReasons']) ? '/reasons' : undefined,
+  'contributory-factors-comments': (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['contributoryFactors']) ? '/contributory-factors' : undefined,
+  'safer-custody': (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['contributoryFactors']) ? '/contributory-factors' : undefined,
+  'additional-information': (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['isSaferCustodyTeamInformed']) ? '/safer-custody' : undefined,
+  'check-answers': (req: Request) =>
+    isMissingValues(req.journeyData.referral!, ['isSaferCustodyTeamInformed']) ? '/safer-custody' : undefined,
 }
